1. 评测记录列表评测方案弹窗数据 2. 评测方案模块数据模型字段对齐 3. 新增biz_models放各个模块的模型避免合并冲突
