+++ dashboard/resources.py
@@ -5,6 +5,7 @@ from typing import List
 from starlette.requests import Request
 
 from dashboard import enums
+from dashboard.biz_models.datamanager import DataSet, EvaluationPlan
 from dashboard.constants import BASE_DIR
 
 # from dashboard.models import Evaluation
@@ -13,7 +14,7 @@ from dashboard.models import Cat  # EvaluationPlan,; Evaluation,
 from dashboard.models import Category  # EvaluationPlan,; Evaluation,
 from dashboard.models import Config  # Evaluation,
 from dashboard.models import Dog1  # EvaluationPlan,; Evaluation,
-from dashboard.models import DataSet, EvaluationPlan, LabelPage, Log  # EvaluationPlan,; Evaluation,
+from dashboard.models import LabelPage, Log  # EvaluationPlan,; Evaluation,
 from dashboard.models import Permission as PermissionModel
 from dashboard.models import Product, Record  # EvaluationPlan,; Evaluation,
 from dashboard.models import Resource as ResourceModel
