+++ dashboard/routes.py
@@ -9,8 +9,9 @@ from starlette.status import HTTP_303_SEE_OTHER, HTTP_404_NOT_FOUND
 from tortoise import Model
 from tortoise.transactions import in_transaction
 
+from dashboard.biz_models.datamanager import DataSet, EvaluationPlan
 from dashboard.biz_routers import biz_router
-from dashboard.models import Config, DataSet, EvaluationPlan, Log, ModelInfo
+from dashboard.models import Config, Log, ModelInfo
 from dashboard.widgets.displays import ShowModelCard
 from last.services.app import app
 from last.services.depends import (
